Rename AddEvaluationForm to AddCategoryForm
