+++ src/app/components/AddCategoryForm/AddCategoryForm.stories.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
-import AddEvaluationForm from './AddEvaluationForm';
+import AddCategoryForm from './AddCategoryForm';
 
 export default {
-  title: 'Component/AddEvaluationForm',
-  component: AddEvaluationForm,
+  title: 'Component/AddCategoryForm',
+  component: AddCategoryForm,
   argTypes: { onSubmit: { action: 'submitted' } },
 };
 
 export const regular = (): JSX.Element => (
-  <AddEvaluationForm
+  <AddCategoryForm
     onSubmit={() => console.log('submitted')}
     missingInput={false}
     onCancel={() => console.log('canceled!')}
