+++ src/app/pages/PupilOverview/PupilOverview.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { useNavigate, useParams } from 'react-router';
 import styled from 'styled-components';
-import AddEvaluationForm from '../../components/AddEvaluationForm/AddEvaluationForm';
+import AddEvaluationForm from '../../components/AddCategoryForm/AddCategoryForm';
 import CopyButton from '../../components/CopyButton/CopyButton';
 import EvaluationCard from '../../components/EvaluationCard/EvaluationCard';
 import Header from '../../components/Header/Header';
